Add interactiveState to embeddables
This currently assumes that the relevant state is in the answer's
`report_state.interactiveState`, which is true for question-interactives,
but may not be true for all answers.

This also deals with an annoying id issue, which might be a bug on
LARA's end, though could be something with the data I'm working with.