Add type definitions for Activity, Page, Embeddable
This adds all the type definitions for the major components of the
exported activity from LARA.

Note that a number of types are, e.g.,

`name?: string | null`

which is ugly, but unfortunately the field may be missing or may be null,
which is of course different in TS.

Note that these types are a first draft. As we import new activities, we
may find we need to make changes, but at least we will be explicit when
we do so.