Move interactiveState out of global activity props
Previously we added the interactive state directly to the individual
embeddables in the global Activity state. Among other issues, this
would make the entire app re-render if the interactive state changed.

This changes things such that

1. app.ts initializes the listener on `answers` as before
2. Individual embeddables request the interactive state data

Since we are watching for changes on `answers`, any embeddable will
always get the freshest state when the embeddable loads. However, once
an embeddable has loaded it will no longer be listening for changes to
the interactive state, as this is not currently supported by the
interactive API.

To make this clearer, we now pass down `initialInteractiveState` to the
nested embeddable components. If this changes (e.g. if we only hear back
from Firestore after the embeddable has already rendered) we will
re-render the embeddable.