Split Embeddable type into different sub-types
This makes some of the code more verbose, as we need to type-guard more
explicitly, but it makes everything more explicit, especially as some of
the embeddable types really don't share a lot of props.

I'm unsure whether all the possible-undefined `prop?:`s are still needed,
as it may be that LARA will always export those props (or `null`) for
those explicit types, but it seems a little flaky what LARA exports so
it seems like it can't hurt to always require checking to see if a prop
exists.