Create embeddable-util for creating Answer objects
This takes an embeddable, a new interactive state, and an optional old
answer object to create a new answer object to be saved to the database.
This new object doesn't yet contain the LTI info which will get added
by `firestore-db`.

This also extracts the refId <-> answer's question-id rewriting code.

Add tests.