Update embeddable so we wait for student data
This changes the embeddable so that we have to wait for the request for
student data to finish before we render it. Until then we will display
"loading..."

Note that we may not have any student data for the embeddable. For this
reason, in order to determine whether any has been requested, firebase-db
has a new flag to check the state of requests. We will set the request
as completed when we get a response from the DB, whether or not there is
data.

The embeddable and iframe-runtime are changed to make it clearer that
the initialInteractiveState can only change once.