Restrict iframe reload to URL changes (#74)
Normally, I’m in favor of following the `exhaustive-deps` rule’s recommendations rather than disabling it, but in this case I think it makes sense to go the other way. With this effect in particular we’ve repeatedly gone through the cycle of adding new code to the effect, adding the recommended identifiers to the list of dependencies, and then figuring out what combination of `useCallback`, `useMemo`, and `useRef` will suffice to make sure that none of them trigger a reload of the `<iframe>`. Rather than perpetuating that cycle, with this PR I propose to say explicitly what we mean, namely that only a URL change should trigger an `<iframe>` reload and any reload triggered otherwise is a bug.