the question ids in the resourceUrl need to match the ids in the contentUrl
a question id mismatch is only seen when looking at the portal-report
what really maters is the `activity` url parameter on the assignment in the portal,
that is what the portal-report uses.
but we'll want the resourceUrl here to match the activity parameter for when
the student starts using the activity when working offline.