hacky untyped fix for the duplicate answers problem
The issue is that the AP code using the storage facade, doesn't update its own answer objects.
It is counting on the storage facade to call its answerWatcher callback and then it updates the
answer object.
So the dexie provider needs to implement the answerWatcher better. The implementation here
does does this watching within the current tab. It isn't using dexie observable or something
similar to watch for changes across tabs.

Also this code isn't supporting the unsubscribing of the watchers. It doesn't look like the AP code
is calling this unsubscribe, in the normal case, but it should be.