This is a WIP of handling the service worker states
The app.tsx is getting big so it'd be great to separate this out.
Right now there is a mix of workbox and native serviceworker apis, hopefully that can be improved.

It also doesn't handle showing the activated state.
We might want to call claimClients once we are activated so we don't need to reload.