handle installing case
if a first time service worker started installing in another tab and hasn't finished when the user opens a new tab,
getSW() will not resolve in the new tab. Also if the user reloads the current tab while it is still installing for the first
time, then getSW() will not resolve. The additions in this commit handle this by looking at _registration.installing
and adding a listener to the serviceWorker if there is one.