switch to cors requests when loading offline manifest
doing this exposed an issue with the existing code. There are two css files referenced
by index.html that are referenced before we start installing the offline manifest
these references were breaking the cache.
my solution is to add them to the pre-cache so they are available before the index.html
request even happens (when the page is controlled).

This also adds the rangerequest plugin which is needed to support videos.
And it removes a broken url from production pc manifest.

I think part of this is that by using cors we are see any download errors. With no-cors
those download errors are cached as if they succeeded so we never knew they failed