Manual rewrite of `simple-carousel.json` to use trailing slashes
This commit manually rewrites the `authored_state` `subinteractives` URL for the `simple-carousel.json` sample activity. The changes add trailing slashes to the subinteractives.  Note that the second page of the carousel fails for other reasons.

With this commit, you should now be able to see the first open response question in the carousel interactive when you open it locally on the online port:

http://127.0.0.1:11000/?activity=offline-activities%2Fsimple-carousel.json