WIP: prototype of new install flow
The user needs to load install.html
Then reload the page which triggers the installation of the application files in the cache.
The user can then go offline and visit index.html and no extra requests should
be made over the network.

There are many thing do still. They are listed as tasks in the story currently:
https://www.pivotaltracker.com/story/show/177164109