skip the disk cache when installing assets
this works around an issue where images are loaded by another
site without a CORS request. The response is stored in the disc
cache. But the key is a simple URL and the value doesn't have
the CORS headers. Now if the installer trys to load that asset it fails
because the response (from the cache) doens't include the corredt
cors headers.