force installs to download all assets from the network
It seems likely we are going to have to change this because it seems very likely
that people are going to have network errors when downloading gigabytes of data
And we wouldn't want to them to have to download all of that data from scratch again.

But this current approach seems safest since the content is still changing.