remove the offline manifest authoring code
We now can generate the manifests statically.
Additionally this code didn't work after the update to using an install page.
With a default CacheOnly Strategy.
Perhaps we'll want to bring this back if we can't statically handle
new activities, but it seems best to remove it since it doesn't currently work.