a new revision to the schema
bump version field to 2

renames
  toggle_info_assessment -> secondary_column_collapsible
  embeddable_display_mode -> secondary_column_display_mode

It removes the extra "embeddable" key which seems redundant.
The only reason I think that was was there, was to keep properties that are stored on the PageItem in lara out of the actual
embeddable object. This distinction is confusing to consumers of this format, so I'm removing it.

changes the column names from to primary and secondary
These names seem better to me since it focuses on the role of the column rather than its size. If we implement
responsiveness, the runtime might combine the two columns into one so it fits better on a small screen. In that
case the secondary won't be smaller anymore.  My only concern with this new approach is the length of the
strings primary and secondary.