Initial Notebook reskin [PT-184609540]
This is the result of the spike to reskin AP to a new "Notebook" layout.  It will be the basis of the real notebook epic work.

This spike includes:

- A force of all activities to use the notebook layout (this will need to change once an authoring option is available)
- A new notebook.scss style file that includes all style changes
- An update of the activity nav to use page names and hide prev/next buttons
- A notebook layout of sections as tabs
- A new sample notebook activity
- A spike of the media library

NOTE: The Cypress tests are disabled in the ci.yml file due to the forcing of all activities to use the notebook layout