feat: Parse media library items from sequence/activity json [PT-185362209]
This parses the media library items from the sequence and activity json.

NOTE: this uses a pre-release of the LARA interactive api client.  There is a chore to update to the final release in the notebook epic.

NOTE: I had to downgrade the engines field in the package json otherwise the peer dependency will fail trying to install the new Lara interactive api as Lara uses a different react version.