feat: Only require header for sharing plugin [PT-187807322]
This updates the code that decides if the header is shown not check if any plugin is associated with the question but only if the sharing plugin is associated.

The code is setup in such a way so that other plugins can be included in that list in the future.