fix: Out of bounds page read in setDocumentTitle [AP-38]
Fixes an error when the showFeedbackPage query parameter is present.  The completion page page number was not being correctly calculated when there were hidden pages and setDocumentTitle was being passed a page number outside the bounds of the page array.

This also adds a sanity check to ensure setDocumentTitle uses a page number within the range of pages to avoid a page crash.